Implement blueprints for detours
