+++ website/templates/home_page.html
@@ -37,7 +37,7 @@
               </li>
               <li>
                 <!-- <i class="ph ph-arrows-left-right"></i> -->
-                <a href="{{ url_for('list_detour') }}">Desvios</a>
+                <a href="{{ url_for('detour_blueprint.list_detour') }}">Desvios</a>
               </li>
               <li>
                 <!-- <i class="ph ph-broadcast"></i> -->
